+++ HISTORY.md
@@ -6,7 +6,7 @@ v0.8 (???/??)
 
 - Fixes:
     - `X.gte`: don't recurse into tasks
-    - `await(ts)`: clear terminated task
+    - `await(ts)`: clear terminated task, clear empty awake
 
 v0.7 (jun/26)
 -------------